requirements.txt
pyserial
discord.py
